+++ v3/src/utilities/boundary-utils.ts
@@ -1,4 +1,27 @@
-import usStateBoundaries from "./US_State_Boundaries.json"
+import usStateBoundaries from "../boundaries/US_State_Boundaries.json"
+
+export const kPolygonNames = ['boundary', 'boundaries', 'polygon', 'polygons', 'grenze', '境界', 'مرز']
+
+export const boundaryObjectFromBoundaryValue = (iBoundaryValue: object | string) => {
+  if (typeof iBoundaryValue === 'object') {
+    return iBoundaryValue
+  } else {
+    try {
+      return JSON.parse(iBoundaryValue)
+    } catch (er) {
+      return null
+    }
+  }
+}
+
+export const isBoundaryValue = (iValue: object | string): boolean => {
+  const obj = boundaryObjectFromBoundaryValue(iValue)
+  return obj != null &&
+    !!(obj.geometry || obj.coordinates || obj.features ||
+      obj.type === 'FeatureCollection' || obj.type === 'Feature' || obj.jsonBoundaryObject)
+}
+
+// Functions to access remote boundary data
 
 function processBoundaries(boundaryDocument: any) {
   const dataset = boundaryDocument.contexts[0]
